Don't force the completion of pending RPC calls on disconnect.
There's nothing sensible to be done there. If the client is simple, they'll be killed with a signal
(that awesomely might race with our force_completion() on the IO thread).

If the client is complex, they'll have a handler for the disconnect.

In neither case is it useful to call the RPC response handlers with a garbage non-reply.
Our response handlers are not generally capable of handling invalid protobuf::Message objects,
so this frequently results in calling a pure-virtual function and termination.

At best this would result in calling a client callback with an error object. But the client
already knows that the connection has died (or the process has been killed by a signal)
so that's not super-useful.