Fix AppMan handling Upstart resuming a Stopped application
AppMan did not expect apps to be restarted by anyone other than itself. However url-dispatcher will launch an instance of an application if it needs to. Should that app be in a Stopped state, AppMan would reject it. This patch has AppMan accept that new process and fire a focus event to the shell.