Add a test to check for proper code formatting.
This treats our .clang-format file as the authoritative source of style guidance,
and checks that clang-format would format the file identically to how it is formatted.

Because this test currently fails everywhere, it only checks explicitly whitelisted
source files in tools/files-to-check-for-formatting. This is a plain text file that contains
a list of filenames, one name per line, relative to the top source dir.