Make dynamic double buffering MUCH more conservative so that it does not accidentally class a client whose performance is borderline as fast (Unity8 dash as experienced in silo testing for Mir 0.14.0).
Previously a client had to perform fast for 3 frames continuously to
qualify. Now it needs to be fast for 100 frames continuously.

This means in reality most input-driven apps won't ever qualify for
double buffering. But we may never fix all our clients adequately, so
err on the side of smoothness with higher latency.