Workaround for GLibMainLoopTest.propagates_exception_from_server_action failure
The test fails with GCC5 when using -O2 options in armhf which seems to be caused by using identical lambda definitions.
As a workaround the exception message is changed to force the compiler to generate a unique implementation instead of sharing one;this allows the exception to be caught as expected.

Cherry-pick from lp:mir r2820