Add DeviceHandle and DeviceSettings
DeviceHandle shall replace InputDeviceInfo within the server api, addtionally to the information it provided up to now, it shall also provide access to the input device settings. Those are represented by DeviceSettings which is a bundle of all allowed settings of a device. Each setting is represented by an id (currently mir::input::DeviceSettings::Setting) and a varint of int, double, string, matrix3x3.