Update mir-client-platform-mesa pkgconfig and package depends.
The mir-client-platform-mesa headers use:
mir_toolkit/mir_native_buffer.h
mir_toolkit/client_types.h

The former is from libmircommon-dev, so add this to the package dependencies.
The latter is from libmirclient-dev so add the appropriate dependency into the
pkgconfig file so the include directories are passed to the compiler.